pyproject.toml
[project]
name = "xyseg"
version = "0.0.1"
description = "Recursive Segmentation Algorithm"
readme = "README.md"
requires-python =">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Johnathan Chiu", email = "johnathanchiu1065@gmail.com"},
]
keywords = ["images", "segmentation", "document", "parser"]
dependencies = ["numpy>=2.1.1", "pdfplumber>=0.11.4", "pillow>=10.4.0"]

[tool.setuptools]
packages = []

[tool.ruff]
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F821", "E402"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["xyseg/**/*.py", "LICENSE", "README.md"]

[project.urls]
Homepage = "https://github.com/johnathanchiu/recursive-segmentation"
Issues = "https://github.com/johnathanchiu/recursive-segmentation/issues"